+++ metatensor-torch/CHANGELOG.md
@@ -8,6 +8,9 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 
 ### Added
 
+- `Labels.difference` and `Labels.difference_and_mapping` for finding the set
+  difference between two `Labels`
+
 - `energy_uncertainty`, `non_conservative_forces` and `non_conservative_stress` are now
   possible outputs for atomistic models